Refactor retention subsystem into `cratedb_toolkit` namespace
