+++ cratedb_toolkit/retention/cli.py
@@ -8,11 +8,11 @@ import click
 from click import ClickException
 from click_aliases import ClickAliasedGroup
 
-from cratedb_retention.core import RetentionJob
-from cratedb_retention.model import JobSettings, RetentionPolicy, RetentionStrategy
-from cratedb_retention.setup.schema import setup_schema
-from cratedb_retention.store import RetentionPolicyStore
 from cratedb_toolkit.model import DatabaseAddress
+from cratedb_toolkit.retention.core import RetentionJob
+from cratedb_toolkit.retention.model import JobSettings, RetentionPolicy, RetentionStrategy
+from cratedb_toolkit.retention.setup.schema import setup_schema
+from cratedb_toolkit.retention.store import RetentionPolicyStore
 from cratedb_toolkit.util.cli import (
     boot_click,
     click_options_from_dataclass,
@@ -139,7 +139,7 @@ tags_option: t.Any = click.option(
 @click.group(cls=ClickAliasedGroup)  # type: ignore[arg-type]
 @click.option("--verbose", is_flag=True, required=False, help="Turn on logging")
 @click.option("--debug", is_flag=True, required=False, help="Turn on logging with debug level")
-@click.version_option(package_name="cratedb-retention")
+@click.version_option()
 @click.pass_context
 def cli(ctx: click.Context, verbose: bool, debug: bool):
     return boot_click(ctx, verbose, debug)
