+++ cratedb_toolkit/retention/model.py
