+++ cratedb_toolkit/retention/setup/schema.sql
