+++ cratedb_toolkit/retention/strategy/reallocate.py
@@ -15,7 +15,7 @@ Prerequisites
 import dataclasses
 import logging
 
-from cratedb_retention.model import RetentionTask
+from cratedb_toolkit.retention.model import RetentionTask
 
 logger = logging.getLogger(__name__)
